requirements.txt
ipython==4.0.1
matplotlib==1.5.0
Pillow==3.0.0
